+++ src/app/kvm/components/KVMForms/AddLxd/AuthenticationForm/AuthenticationForm.test.tsx
