+++ src/app/kvm/components/KVMForms/ComposeForm/InterfacesTable/SubnetSelect/SubnetSelect.test.tsx
