+++ src/app/kvm/components/KVMForms/ComposeForm/InterfacesTable/SubnetSelect/SubnetSelect.tsx
