+++ src/app/kvm/components/KVMForms/DeleteForm/DeleteForm.tsx
