+++ src/app/kvm/components/KVMForms/DeleteForm/index.ts
